move xdr specific functions to xdr_encoder.go
